:truck: divide view tests
